feat: add llm filter interface
